Change the step section of bakeru file
I added new steps that was missing, and edit steps that uses the refference of the step orders. I realized the I did not add 2 combination steps, so I added them back as step 1 and step 2, which messes up the ordianry step orders and the order step reference inside it. So I edit them again to match the orders correctly.